+++ src/mod-compatibility/all.ts
@@ -1,3 +1,4 @@
 import './nax-art-switch'
 import './lqm-joern-mod'
 import './jetpack'
+import './cc-map-screenshot'